requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
google-cloud-firestore
